worker/requirements.txt
aio_pika
psycopg2-binary
boto3
pydantic
pyyaml
